refactor: move errors.ts
