+++ src/error/errors.ts
@@ -1,7 +1,7 @@
 import path from "path";
 import { cwd } from "process";
-import { AstFuncId, AstId, AstTypeId } from "./grammar/ast";
-import { SrcInfo } from "./grammar";
+import { AstFuncId, AstId, AstTypeId } from "../grammar/ast";
+import { SrcInfo } from "../grammar";
 
 export class TactError extends Error {
     readonly loc?: SrcInfo;
