+++ src/grammar/prev/parser-error.ts
@@ -1,6 +1,6 @@
 import { MatchResult } from "ohm-js";
 import { ErrorDisplay } from "../../error/display";
-import { TactCompilationError } from "../../errors";
+import { TactCompilationError } from "../../error/errors";
 import { syntaxErrorSchema } from "../parser-error";
 import { ItemOrigin, SrcInfo } from "../src-info";
 import { getSrcInfoFromOhm } from "./src-info";
