+++ src/optimizer/test/partial-eval.spec.ts
@@ -13,7 +13,7 @@ import { AssociativeRule3 } from "../associative";
 import { evalBinaryOp, evalUnaryOp } from "../../interpreter";
 import { getParser } from "../../grammar";
 import { defaultParser } from "../../grammar/grammar";
-import { throwInternalCompilerError } from "../../errors";
+import { throwInternalCompilerError } from "../../error/errors";
 
 const MAX: string =
     "115792089237316195423570985008687907853269984665640564039457584007913129639935";
