+++ src/types/resolveStatements.ts
@@ -17,7 +17,7 @@ import {
     throwCompilationError,
     throwConstEvalError,
     throwInternalCompilerError,
-} from "../errors";
+} from "../error/errors";
 import {
     getAllStaticFunctions,
     getStaticConstant,
